pinaka_django/requirements.txt
asgiref==3.11.0
attrs==25.4.0
Django==4.2
django-cors-headers==4.3.0
django-environ==0.10.0
djangorestframework==3.14.0
drf-spectacular==0.27.0
inflection==0.5.1
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
Pillow==11.3.0
psycopg2-binary==2.9
pytz==2025.2
PyYAML==6.0.3
referencing==0.36.2
rpds-py==0.27.1
sqlparse==0.5.3
typing_extensions==4.15.0
uritemplate==4.2.0
